+++ portal-impl/src/com/liferay/portlet/internal/PortletContextImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import aQute.bnd.annotation.ProviderType;
 
@@ -26,6 +26,7 @@ import com.liferay.portal.kernel.portlet.LiferayPortletContext;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.security.lang.DoPrivilegedUtil;
+import com.liferay.portlet.PortletRequestDispatcherImpl;
 
 import java.io.InputStream;
