+++ portal-impl/test/unit/com/liferay/portlet/PortletRequestDispatcherImplTest.java
@@ -24,6 +24,7 @@ import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.model.impl.PortletAppImpl;
 import com.liferay.portal.model.impl.PortletImpl;
 import com.liferay.portal.util.PortalImpl;
+import com.liferay.portlet.internal.PortletContextImpl;
 import com.liferay.registry.BasicRegistryImpl;
 import com.liferay.registry.RegistryUtil;
